% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylationExperiment-Constructor.R
\docType{methods}
\name{MethylationExperiment}
\alias{MethylationExperiment}
\title{High level constructor for MethylationExperiment class}
\usage{
MethylationExperiment(ClassName, ...)
}
\description{
This is a higher level function that allows to construct
   a MethylationExperimentLimma or MethylationExperimentBeta object with
   the colData and rawData.
}
