% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylationExperimentBeta-Class.R,
%   R/MethylationExperimentLimma-Class.R
\docType{class}
\name{MethylationExperimentLimma-class}
\alias{MethylationExperimentLimma-class}
\alias{.MethylationExperimentBeta}
\alias{.MethylationExperimentLimma}
\title{MethylationExperimentBeta-Class}
\description{
This is a concrete class extends MethylationExperiment to
   impose a beta distribution using betareg package to fit it. In this
   context, as it inherits from a SummarizedExperiment the fitted models
   will be included a "model" column, in the features (rowData), where each
   model is stored as a list.

This is a concrete class extends MethylationExperiment to impose a normal
   distribution using limma package to fit it. In this context, as it
   inherits from a SummarizedExperiment the fitted models will be included
   a "model" column, in the features (rowData), where each model is stored
   as a list.
}
